+++ src/Epinova.ElasticSearch.Core.EPiServer/Providers/ElasticSearchMenuProvider.cs
@@ -12,7 +12,7 @@ namespace Epinova.ElasticSearch.Core.EPiServer.Providers
     {
         private static Func<RequestContext, bool> GetAccessInfo()
         {
-            return request => PrincipalInfo.CurrentPrincipal.IsInRole("ElasticsearchAdmins");
+            return request => PrincipalInfo.CurrentPrincipal.IsInRole(RoleNames.ElasticsearchAdmins);
         }
 
         private readonly Func<string, string> _translate = key => LocalizationService.Current.GetString(String.Concat("/epinovaelasticsearch/", key));