+++ src/Epinova.ElasticSearch.Core.EPiServer/RoleNames.cs
@@ -1,0 +1,8 @@
+﻿namespace Epinova.ElasticSearch.Core.EPiServer
+{
+    internal static class RoleNames
+    {
+        internal const string ElasticsearchAdmins = nameof(ElasticsearchAdmins);
+        internal const string ElasticsearchEditors = nameof(ElasticsearchEditors);
+    }
+}