+++ src/Epinova.ElasticSearch.Core/Admin/Index.cs
@@ -110,6 +110,24 @@ namespace Epinova.ElasticSearch.Core.Admin
                 return false;
             }
         }
+        
+        internal int GetDocumentCount()
+        {
+            var uri = _indexing.GetUri(_name, "_search") + "?size=0";
+            dynamic model = new { hits = new { total = 0 } };
+            
+            try
+            {
+                string response = HttpClientHelper.GetString(new Uri(uri));
+                var result = JsonConvert.DeserializeAnonymousType(response, model);
+                return result.hits.total;
+            }
+            catch (Exception ex)
+            {
+                Logger.Error("Could not get count", ex);
+                return 0;
+            }
+        }
 
         internal void Initialize(Type type)
         {
@@ -122,6 +140,7 @@ namespace Epinova.ElasticSearch.Core.Admin
             _indexing.CreateIndex(_name);
 
             CreateStandardSettings();
+            CreateAttachmentPipeline();
 
             _indexing.Close(_name);
 
@@ -206,6 +225,18 @@ namespace Epinova.ElasticSearch.Core.Admin
             HttpClientHelper.Put(uri, data);
         }
 
+        private void CreateAttachmentPipeline()
+        {
+            string json = Serialization.Serialize(Pipelines.Attachment.Definition);
+            byte[] data = Encoding.UTF8.GetBytes(json);
+            var uri = new Uri($"{_settings.Host}/_ingest/pipeline/{Pipelines.Attachment.Name}");
+
+            Logger.Information("Creating Attachment Pipeline");
+            Logger.Information(JToken.Parse(json).ToString(Formatting.Indented));
+
+            HttpClientHelper.Put(uri, data);
+        }
+
         private void CreateTriGramTokenizer()
         {
             string json = Serialization.Serialize(Analyzers.TriGramTokenizer);
@@ -237,7 +268,7 @@ namespace Epinova.ElasticSearch.Core.Admin
             var uri = _indexing.GetUri(_name, "_settings");
             HttpClientHelper.Put(uri, data);
 
-            Logger.Information($"Adding raw analyzer:\n{json}");
+            Logger.Information($"Adding raw analyzer:\n{JToken.Parse(json).ToString(Formatting.Indented)}");
         }
 
         private void CreateSuggestAnalyzer()