+++ src/Epinova.ElasticSearch.Core/Models/Query/MatchBase.cs
@@ -1,4 +1,5 @@
 namespace Epinova.ElasticSearch.Core.Models.Query
 {
+    //TODO: Should this be an interface? Check IsAssignableFrom in RangeConverter if converting
     internal abstract class MatchBase { }
 }