+++ changelog.md
@@ -1,7 +1,9 @@
 # Changelog
 
 #### 10.3
-
+* ES 6 support
+* Replace Attachment Mapper with Ingest plugin
+* No dynamic mapping, base on analyzed content
 * Commerce overhaul, catalog content now has its own index
 * MoreLikeThis function including widget
 * Moved to GitHub